+++ test/fetch.js
@@ -1,7 +1,14 @@
-global.self = {}
-try { require("whatwg-fetch"); module.exports = global.self.fetch }
+try {
+  var self = global.self = {}
+  require("whatwg-fetch")
+  exports = module.exports = self.fetch
+  exports.Headers = self.Headers
+  exports.Request = self.Request
+  exports.Response = self.Response
+
+}
 finally { delete global.self }
 
 // Cannot reload the polyfill as it mutates world state and doesn't return
 // exports.
-require("require-guard")()
+require("require-guard")(module.id)