+++ test/index_test.js
@@ -1,7 +1,8 @@
 var Sinon = require("sinon")
-var HttpError = require("standard-http-error")
+var Fetch = require("./fetch")
+var FetchError = require("../fetch_error")
 var ErrorifyFetch = require("..")
-var fetch = ErrorifyFetch(require("./fetch"))
+var fetch = ErrorifyFetch(Fetch)
 
 describe("ErrorifyFetch", function() {
   beforeEach(function() {
@@ -20,23 +21,48 @@ describe("ErrorifyFetch", function() {
     ;(yield res.text()).must.equal("Hello")
   })
 
-  it("must reject with HttpError if 400 Bad Request", function*() {
+  it("must reject with FetchError if 400 Bad Request", function*() {
     var res = fetch("/")
     this.requests[0].respond(400, {}, "")
 
     var err
     try { yield res } catch (ex) { err = ex }
-    err.must.be.an.instanceof(HttpError)
+    err.must.be.an.instanceof(FetchError)
     err.code.must.equal(400)
+    err.message.must.equal("Bad Request")
   })
 
-  it("must reject with HttpError if 500 Internal Server Error", function*() {
+  it("must reject with FetchError if 500 Internal Server Error", function*() {
     var res = fetch("/")
     this.requests[0].respond(500, {}, "")
 
     var err
     try { yield res } catch (ex) { err = ex }
-    err.must.be.an.instanceof(HttpError)
+    err.must.be.an.instanceof(FetchError)
     err.code.must.equal(500)
+    err.message.must.equal("Internal Server Error")
+  })
+
+  it("must assign response on FetchError if a non-OK response", function*() {
+    var res = fetch("/nonexistent")
+    this.requests[0].respond(404, {}, "Lost it. :(")
+
+    var err
+    try { yield res } catch (ex) { err = ex }
+    err.must.have.nonenumerable("response")
+    err.response.must.be.an.instanceof(Fetch.Response)
+    err.response.status.must.equal(404)
+  })
+
+  it("must assign request on FetchError if a non-OK response", function*() {
+    var res = fetch("/nonexistent", {headers: {"Accept": "application/vnd.x"}})
+    this.requests[0].respond(404, {}, "Lost it. :(")
+
+    var err
+    try { yield res } catch (ex) { err = ex }
+    err.must.have.nonenumerable("request")
+    err.request.must.be.an.instanceof(Fetch.Request)
+    err.request.url.must.equal("/nonexistent")
+    err.request.headers.get("Accept").must.equal("application/vnd.x")
   })
 })