+++ CHANGELOG.md
@@ -1,2 +1,9 @@
+## Unreleased
+- Rejects with `FetchError` instead of `HttpError`.
+- Adds the `request` property (an instance of [`Request`][]) with details of the
+  request to the error.
+
+[`Request`]: https://developer.mozilla.org/en-US/docs/Web/API/Request
+
 ## 0.1.337 (Aug 14, 2015)
 - Scoped release for personal use and profit.